+++ test/test_SOAPify.py
@@ -55,12 +55,13 @@ def test_concatenationOfSOAPreferences():
     assert_array_equal(b.spectra[0], c.spectra[2])
     assert_array_equal(b.spectra[1], c.spectra[3])
 
+
 def test_concatenationOfSOAPreferencesLonger():
     a = SOAPReferences(["a", "b"], [[0, 0], [1, 1]], 8, 8)
     b = SOAPReferences(["c", "d"], [[2, 2], [3, 3]], 8, 8)
     d = SOAPReferences(["e", "f"], [[4, 4], [5, 5]], 8, 8)
-    c = SOAPify.mergeReferences(a, b,d)
-    assert len(c.names) == (len(a)+len(b)+len(d))
+    c = SOAPify.mergeReferences(a, b, d)
+    assert len(c.names) == (len(a) + len(b) + len(d))
     assert c.spectra.shape == (6, 2)
     assert_array_equal(a.spectra[0], c.spectra[0])
     assert_array_equal(a.spectra[1], c.spectra[1])
@@ -68,6 +69,7 @@ def test_concatenationOfSOAPreferencesLonger():
     assert_array_equal(b.spectra[1], c.spectra[3])
     assert_array_equal(d.spectra[0], c.spectra[4])
     assert_array_equal(d.spectra[1], c.spectra[5])
+
 
 def test_fillSOAPVectorFromdscribeSingleVector():
     nmax = 4
@@ -101,3 +103,60 @@ def test_fillSOAPVectorFromdscribeArrayOfVector():
                     assert c[l, n, np] == a[i, limited]
                     assert c[l, np, n] == a[i, limited]
                     limited += 1
+
+
+def test_transitionMatrixWithNoError():
+    data = SOAPify.SOAPclassification(
+        [],
+        numpy.array(
+            [
+                [0, 0, 1, 2, 2, 1],
+                [0, 0, 2, 2, 2, 2],
+                [0, 0, 2, 2, 2, 1],
+                [0, 0, 2, 2, 2, 2],
+            ]
+        ),
+        ["1", "2", "3"],
+    )
+    tmat = SOAPify.transitionMatrixFromSOAPClassification(data)
+    assert tmat.shape[0] == len(data.legend)
+    # hand calculated:
+    assert_array_equal(
+        tmat,
+        numpy.array(
+            [
+                [6, 0, 0],
+                [0, 0, 3],
+                [0, 1, 8],
+            ]
+        ),
+    )
+
+
+def test_transitionMatrixWithError():
+    data = SOAPify.SOAPclassification(
+        [],
+        numpy.array(
+            [
+                [0, 0, 1, 2, 2, 1],
+                [0, 0, 2, 2, 2, 2],
+                [0, 0, 2, 2, 2, 1],
+                [0, 0, 2, 2, 2, -1],
+            ]
+        ),
+        ["1", "2", "3", "Errors"],
+    )
+    tmat = SOAPify.transitionMatrixFromSOAPClassification(data)
+    assert tmat.shape[0] == len(data.legend)
+    # hand calculated:
+    assert_array_equal(
+        tmat,
+        numpy.array(
+            [
+                [6, 0, 0, 0],
+                [0, 0, 2, 1],
+                [0, 1, 8, 0],
+                [0, 0, 0, 0],
+            ]
+        ),
+    )