+++ Changelog.md
@@ -5,6 +5,7 @@
 - `getXYZfromTrajGroup` now accept IO objects as inputs
 - **WARNING**: broken interface for `getXYZfromTrajGroup`, now it needs 3 inputs and the first is a file-like object
 - `saveXYZfromTrajGroup` and `getXYZfromTrajGroup` now can export comments per frames
+- `transitionMatrixFromSOAPClassification` now creates matrix with shape  `(n,n)` and no more `(n+1,n+1)`, where `n` is the lenght of the legend. The user will now need to address the errors in classification, if needed
 
 ## Changes From 0.0.2a